requirements.txt
aiohttp==3.7.4.post0
arxiv==1.4.2
async-timeout==3.0.1
attrs==22.1.0
chardet==4.0.0
discord.py==1.7.3
feedparser==6.0.10
idna==3.3
multidict==6.0.2
sgmllib3k==1.0.0
typing_extensions==4.3.0
yarl==1.7.2
pyyaml==5.3.1
psycopg2==2.9.3